requirements.txt
matplotlib==3.8.2
python-bidi==0.4.2
arabic-reshaper==3.0.0
Pillow==10.1.0
ttkbootstrap==1.10.1
